requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
pika>=1.3.0
setuptools
psutil>=5.9.0

